+++ themes/alpha/assets/alpha/liva/js/modules/constants.js
@@ -1,0 +1,26 @@
+// /assets/alpha/liva/js/modules/constants.js
+
+export const LivaPositions = {
+  BOTTOM_LEFT: "liva-bottom-left",
+  BOTTOM_RIGHT: "liva-bottom-right",
+};
+
+export const LivaStates = {
+  HIDDEN: "liva-hidden",
+  ERROR: "liva-fab-error",
+};
+
+export const ApiUrls = {
+  HUGO_VERSION: "https://api.github.com/repos/gohugoio/hugo/tags",
+  THEME_VERSION: "https://alpha.oxypteros.com/alpha/version.json",
+};
+
+export const DocUrls = {
+  ALPHA_UPDATE: "https://alpha.oxypteros.com/update",
+  HUGO_UPDATE: "https://alpha.oxypteros.com/hugo/update",
+};
+
+export const StorageKeys = {
+  ERROR_LIST: "liva-error-list",
+  FAB_POSITION: "liva-pos",
+};