+++ themes/alpha/assets/alpha/liva/js/modules/dom.js
@@ -1,0 +1,57 @@
+// /assets/alpha/liva/js/modules/dom.js
+
+export const livaContainer = document.querySelector(
+  '[data-alpha="liva-container"]',
+);
+export const livaConsoleContainer = document.querySelector(
+  '[data-alpha="liva-console-container"]',
+);
+export const livaConsoleContent = document.querySelector(
+  '[data-alpha="liva-console-content"]',
+);
+
+// FAB
+export const livaFab = document.querySelector('[data-alpha="liva-fab-button"]');
+
+// Settings
+export const livaSettings = document.querySelector(
+  '[data-alpha="liva-settings"]',
+);
+export const livaSettingsCloseBtn = document.querySelector(
+  '[data-alpha="liva-close-settings"]',
+);
+export const selectPositionElement = document.querySelector(
+  '[data-alpha="liva-select-position"]',
+);
+export const livaOpenConsoleBtn = document.querySelector(
+  '[data-alpha="liva-open-console-button"]',
+);
+export const livaUpdateBtn = document.querySelector(
+  '[data-alpha="liva-updates-button"]',
+);
+
+// Console
+export const livaConsole = document.querySelector('[data-alpha="liva-console"]');
+export const livaConsoleCloseBtn = document.querySelector(
+  '[data-alpha="liva-close-console"]',
+);
+export const livaConsoleHelpBtn = document.querySelector(
+  '[data-alpha="liva-console-help-button"]',
+);
+export const livaConsoleResetBtn = document.querySelector(
+  '[data-alpha="liva-console-reset-button"]',
+);
+
+// Version Info
+export const localAlphaVersionEl = document.querySelector(
+  '[data-alpha="local-alpha-version"]',
+);
+export const localHugoVersionEl = document.querySelector(
+  '[data-alpha="local-hugo-version"]',
+);
+export const alphaBadge = document.querySelector(
+  '[data-alpha="liva-alpha-update-badge"]',
+);
+export const hugoBadge = document.querySelector(
+  '[data-alpha="liva-hugo-update-badge"]',
+);