+++ themes/alpha/assets/js/goatcounter.js
@@ -1,16 +1,19 @@
 (() => {
   if (localStorage.getItem("tracking") === "false") return;
-	if (navigator.doNotTrack === "1" && localStorage.getItem("tracking") === null) {
-		localStorage.setItem("tracking", "false"); 
-		return;
-	}
-  (() => {
-    // GoatCounter: https://www.goatcounter.com
-    // This file is released under the ISC license: https://opensource.org/licenses/ISC
-    if (window.goatcounter && window.goatcounter.vars)
-      // Compatibility with very old version; do not use.
-      window.goatcounter = window.goatcounter.vars;
-    else window.goatcounter = window.goatcounter || {};
+  if (
+    navigator.doNotTrack === "1" &&
+    localStorage.getItem("tracking") === null
+  ) {
+    localStorage.setItem("tracking", "false");
+    return;
+  }
+
+  // GoatCounter: https://www.goatcounter.com
+  // This file is released under the ISC license: https://opensource.org/licenses/ISC
+  (function () {
+    "use strict";
+
+    window.goatcounter = window.goatcounter || {};
 
     // Load settings from data-goatcounter-settings.
     var s = document.querySelector("script[data-goatcounter]");
@@ -39,7 +42,8 @@
     var enc = encodeURIComponent;
 
     // Get all data we're going to send off to the counter endpoint.
-    var get_data = function (vars) {
+    window.goatcounter.get_data = function (vars) {
+      vars = vars || {};
       var data = {
         p: vars.path === undefined ? goatcounter.path : vars.path,
         r: vars.referrer === undefined ? goatcounter.referrer : vars.referrer,
@@ -114,8 +118,9 @@
     // Get the endpoint to send requests to.
     var get_endpoint = function () {
       var s = document.querySelector("script[data-goatcounter]");
-      if (s && s.dataset.goatcounter) return s.dataset.goatcounter;
-      return goatcounter.endpoint || window.counter; // counter is for compat; don't use.
+      return s && s.dataset.goatcounter
+        ? s.dataset.goatcounter
+        : goatcounter.endpoint;
     };
 
     // Get current path.
@@ -149,7 +154,7 @@
     };
 
     // Filter some requests that we (probably) don't want to count.
-    goatcounter.filter = function () {
+    window.goatcounter.filter = function () {
       if (
         "visibilityState" in document &&
         document.visibilityState === "prerender"
@@ -173,7 +178,7 @@
 
     // Get URL to send to GoatCounter.
     window.goatcounter.url = function (vars) {
-      var data = get_data(vars || {});
+      var data = window.goatcounter.get_data(vars || {});
       if (data.p === null)
         // null from user callback.
         return;
@@ -287,7 +292,10 @@
         for (var k in opt.attr) d.setAttribute(k, opt.attr[k]);
 
         var p = document.querySelector(opt.append);
-        if (!p) return warn("visit_count: append not found: " + opt.append);
+        if (!p)
+          return warn(
+            "visit_count: element to append to not found: " + opt.append,
+          );
         p.appendChild(d);
       });
     };